pyproject.toml
[project]
name = "django-llm-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "django-constance>=4.3.2",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-community>=0.3.28",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "uuid>=1.30",
]
